refactor: v1.4 post-release Batch A+B (bus/cpu architecture cleanup)
Implements the two batches from docs/internal/v1.4_post_release_optimization_plan.md.

Batch A (ctest 维护期 — mechanical, no roadmap dependency):
  §1.2 src/bus.cpp — set_read_handler / set_write_handler: ascending
    loop with end < start early-out; replaces descending loop that
    underflowed on bad input.
  §2.3 src/bus.cpp — Bus::Bus() ctor: 22 per-array memsets consolidated
    to a single whole-object memset(this, 0, sizeof(*this)).
  §2.4 src/bus.h — inline forwarders setmirror(int) / setmirrorw(int,
    int, int, int) retyped to uint32_t to match Bus::setmirror(uint32_t).

Batch B (v1.5 Prism 启动前 — owes back v1.3 Legion + prepares v1.5):
  §1.1 src/fceu.cpp — fceu11::g_bus.init() replaces the legacy
    SetReadHandler(0, 0xFFFF, ANull) / SetWriteHandler(0, 0xFFFF, BNull)
    pair at the top of PowerNES(). Also fixes a plan bug: init() no
    longer calls reset_mapping() (which would wipe prg_ptr_[] /
    chr_ptr_[] populated by FCEUXLoad before PowerNES).
  §1.3 src/cpu.{h,cpp} — extern uint64_t timestampbase; hoisted from
    cpu.cpp function body to cpu.h, grouped with the other CPU-state
    externs.
  §1.4 src/cpu.{h,cpp} — three std::function dead members
    cpu_hook_/read_hook_/write_hook_ and their set_*_hook setters
    removed (Roadmap §3.1 promise never landed). <functional> include
    dropped.
  §2.1 src/{cpu,fceu,ppu,x6502}.{h,cpp} — bool ::overclocking migrated
    into fceu11::Cpu as a private member with public accessors
    (Cpu::overclocking() / Cpu::set_overclocking()). x6502.cpp reads
    via g_cpu.overclocking(); ppu.cpp writes via g_cpu.set_overclocking().
    Honors the v1.3 Legion §3.2 list (which had missed this one).
  §2.2 src/cpu.{h,cpp} + src/bus.cpp — Cpu::db() / set_db() / pi() /
    set_pi() accessors added (plus set_a/set_x/set_y/set_s/set_p for
    symmetry). Bus::ANullImpl now uses g_cpu.db() instead of reaching
    through native_layout().
  §2.5 src/bus.cpp — v1.5 Prism TODO comment added above the Bus
    class body, marking the PPUCHRRAM / PPUNTARAM / vnapage[] writes
    in setchr*/setmirror*/setntamem as the natural seam where v1.5
    will pull.

Validation:
  ctest -C Release -LE perf: 17/17 PASS (functional gate green)
  bench_tolerance_test (advisory per ec3c2dc): +5-10% on PPU/Full
  golden_savestate_test --compare-layout: 7/7 layout-identical

The bench regression is the same link-time code layout disturbance
root cause as Phase 6 (vrc7_PreSave init-time call): Batch A + B
removes 22 memsets + 3 std::function members + several accessors,
shrinking the text section and shifting hot-path alignment. Per the
post-release plan's §6 the bench is gated only on ctest, and per
ec3c2dc bench_tolerance_test is advisory on shared runners.
Functional and golden invariants are bit-exact preserved.

Performance recovery is scheduled for v1.14 Anvil §14.3 (MSVC /LTCG
+ PGO) per docs/v1.x_Modernization_Roadmap.md. No re-baseline needed
until v1.5 Ppu class lands (text section will jump again).

+++ src/bus.h
@@ -282,12 +282,16 @@ inline void setprg32(uint32_t A, uint32_t V) noexcept { fceu11::g_bus.setprg32(A
 inline void setchr1 (uint32_t A, uint32_t V) noexcept { fceu11::g_bus.setchr1 (A, V); }
 inline void setchr4 (uint32_t A, uint32_t V) noexcept { fceu11::g_bus.setchr4 (A, V); }
 inline void setchr8 (uint32_t V)          noexcept { fceu11::g_bus.setchr8 (V); }
-inline void setmirror (int t)              noexcept { fceu11::g_bus.setmirror (static_cast<uint32_t>(t)); }
-inline void setmirrorw(int a, int b, int c, int d) noexcept {
-    fceu11::g_bus.setmirrorw(static_cast<uint32_t>(a),
-                             static_cast<uint32_t>(b),
-                             static_cast<uint32_t>(c),
-                             static_cast<uint32_t>(d));
+// v1.4 Post-Release Optimization Plan §2.4 — forwarder params
+// match Bus::setmirror(uint32_t) / Bus::setmirrorw(uint32_t, ...).
+// A grep of src/boards/ for setmirror( / setmirrorw( shows all
+// callers pass either MI_0..MI_3 macros (uint32 literals) or
+// uint8 / int local variables — uint8→uint32_t and int→uint32_t
+// implicit conversions are well-defined, so no cast is needed
+// at the call site.
+inline void setmirror (uint32_t t)              noexcept { fceu11::g_bus.setmirror (t); }
+inline void setmirrorw(uint32_t a, uint32_t b, uint32_t c, uint32_t d) noexcept {
+    fceu11::g_bus.setmirrorw(a, b, c, d);
 }
 inline void setntamem(uint8_t* p, int ram, uint32_t b) noexcept {
     fceu11::g_bus.setntamem(p, ram, b);
